Refine abusive voter detection and handling
This update improved the handling of abusive voting behavior. Changes were made to clearly define the number of votes a user can make within a certain time frame before being considered abusive. We now better monitor and modify the cooldown data for voters, and handle situations when the threshold for abusive voting is reached.